requirements.txt
Django==2.2.11
django-templated-mail==1.1.1
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
djoser==1.1.5
drf-nested-routers==0.90.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.3
python-dotenv==0.8.2
pytz==2018.4
